Remove unused download exports (#87)
* Import from src, not lib.

* Remove unnecessary exports.

* Move resource loader to where it's actually used.

* Remove unnecessary exports.

* Move command to index file.

* Enforce config file cannot be a path.

* Fix tests.
