+++ src/index.ts
@@ -5,24 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export { FatRamlResourceLoader } from "./download";
-
 export { getBearer } from "./download/bearerToken";
-export {
-  searchExchange,
-  downloadRestApi,
-  downloadRestApis,
-  getVersionByDeployment,
-  getSpecificApi,
-  getAsset
-} from "./download/exchangeDownloader";
-export {
-  groupByCategory,
-  removeVersionSpecificInformation,
-  removeRamlLinks
-} from "./download/exchangeTools";
-export { extractFiles } from "./download/exchangeDirectoryParser";
-
 export { RestApi } from "./download/exchangeTypes";
 export { ramlToolLogger } from "./common/logger";
